finnished voyages to the point that you can now perform them change them itd.

+++ harbour_heven/lib/data/model/voyage/voyage.dart
@@ -5,7 +5,7 @@ import 'package:harbour_heven/data/model/enum/recource_type.dart';
 class Voyage {
   final Difficulty difficulty;
   final Map<Recourcetype,int> recources;
-  final double successThreshold;
+  final int successThreshold;
 
   Voyage({
     required this.difficulty,
